requirements.txt
flask
flask-socketio
eventlet
python-dotenv
flask-caching
requests
werkzeug
gunicorn
psycopg2-binary
Pillow
schedule

# PostgreSQL Database
psycopg2-binary==2.9.10

# Core Flask dependencies
Flask==3.1.0
Flask-SocketIO==5.5.0
Flask-Caching==2.3.1
Werkzeug==3.1.3
Jinja2==3.1.5
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8

# SocketIO dependencies
python-socketio==5.12.0
python-engineio==4.11.1
bidict==0.23.1
simple-websocket==1.1.0

# Security and encryption
cryptography==44.0.0
cffi==1.17.1
pycparser==2.22

# Environment and configuration
python-dotenv==1.1.0

# Caching (Redis optional)
redis==6.2.0
cachelib==0.13.0

# Asset minification
cssmin==0.2.0
jsmin==3.0.1

# File watching for development (optional)
watchdog==4.0.1

# HTTP requests
requests==2.32.3
urllib3==2.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10

# Server options
gunicorn==23.0.0
gevent==25.5.1
gevent-websocket==0.10.1
greenlet==3.2.3
waitress==3.0.2

# Development and testing
pytest==8.3.5
pluggy==1.5.0
iniconfig==2.1.0
packaging==24.2

# IoT device support
tinytuya==1.15.1

# Utilities and misc
jsonify==0.5
blinker==1.9.0
colorama==0.4.6
h11==0.14.0
wsproto==1.2.0
zope.event==5.1
zope.interface==7.2